Split VM extensions class into smaller classes specific to VM.
