+++ src/NuGetGallery/Helpers/ViewModelExtensions/DisplayLicenseViewModelExtensions.cs
@@ -1,0 +1,39 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Collections.Generic;
+using System.Linq;
+using NuGet.Services.Entities;
+using NuGet.Services.Licenses;
+
+namespace NuGetGallery
+{
+    public static class DisplayLicenseViewModelExtensions
+    {
+        public static DisplayLicenseViewModel Setup(
+            this DisplayLicenseViewModel viewModel,
+            Package package,
+            IReadOnlyCollection<CompositeLicenseExpressionSegment> licenseExpressionSegments,
+            string licenseFileContents)
+        {
+            ((PackageViewModel)viewModel).Setup(package);
+
+            viewModel.EmbeddedLicenseType = package.EmbeddedLicenseType;
+            viewModel.LicenseExpression = package.LicenseExpression;
+            if (PackageHelper.TryPrepareUrlForRendering(package.LicenseUrl, out string licenseUrl))
+            {
+                viewModel.LicenseUrl = licenseUrl;
+
+                var licenseNames = package.LicenseNames;
+                if (!string.IsNullOrEmpty(licenseNames))
+                {
+                    viewModel.LicenseNames = licenseNames.Split(',').Select(l => l.Trim()).ToList();
+                }
+            }
+            viewModel.LicenseExpressionSegments = licenseExpressionSegments;
+            viewModel.LicenseFileContents = licenseFileContents;
+
+            return viewModel;
+        }
+    }
+}
